+++ app/src/main/java/com/example/fuellog/models/Transport.kt
@@ -9,10 +9,22 @@ import androidx.room.PrimaryKey
  * @Date: 2025/12/02
  */
 
+/**
+ * Entity for save Transport
+ * Using DB Room
+ *
+ * @property id Long, Unique record identifier (automatically generated)
+ * @property name String, name of transport
+ * @property company String, company name of transport
+ * @property model String, name model of transport
+ * @property year Int, year made of transport
+ * @property description String, description, information, notes of transport
+ */
+
 @Entity
 data class Transport(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo var id: Int,
+    @ColumnInfo var id: Long,
     @ColumnInfo var name: String?,
     @ColumnInfo var company: String?,
     @ColumnInfo var model: String?,
